+++ src/__tests__/Navbar.test.js
@@ -1,14 +1,18 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
+import { Provider } from 'react-redux';
+import store from '../redux/store';
 import { MemoryRouter as Router } from 'react-router-dom';
 import NavBar from '../components/NavBar';
 
 it('renders correctly', () => {
   const tree = renderer
     .create(
-      <Router>
-        <NavBar />
-      </Router>,
+      <Provider store={store}>
+        <Router>
+          <NavBar />
+        </Router>
+      </Provider>
     )
     .toJSON();
   expect(tree).toMatchSnapshot();
